+++ axelor-supplychain/src/main/java/com/axelor/apps/supplychain/exception/IExceptionMessage.java
@@ -195,6 +195,9 @@ public interface IExceptionMessage {
   static final String SALE_ORDER_LINE_AVAILABILITY_REQUEST = /*$$(*/
       "The reservation for an availability requested stock move cannot be lowered." /*)*/;
 
+  static final String SALE_ORDER_LINE_REQUESTED_QTY_TOO_LOW = /*$$(*/
+      "The requested quantity must be greater than the already delivered quantity." /*)*/;
+
   /** Account config supplychain service */
   static final String FORECASTED_INVOICE_CUSTOMER_ACCOUNT = /*$$(*/
       "You must configure a forecasted invoiced customer account for the company %s" /*)*/;